refactor(dashboard): Import UserDto from one path across the feature
Changes:
- Point admin-activity-card and user-activity at @/lib/api/schemas/auth-user

The dashboard was split two-vs-two: the newer WAU/MAU files went through the
@/lib/api/types barrel while the user-table files imported the schema module
directly. Both resolve to the same type, but one feature folder using two
paths for one type is churn waiting to happen.

Notes:
- Review finding 40 (reviews/REVIEW-2026-07-24-DEV-VS-MAIN.md)
- Only the dashboard is normalised here. The contact-side files also use the
  barrel, but they are outside this review's scope.
